#23 Android, integration od new argument in method init, and fix exception when we open database
